feature/tailnetlock/tslockjsonv1: extract JSON marshalers for `tailscale lock`
This patch extracts the functions used to marshal the JSON output of
the `tailscale lock` subcommand.

Updates #17613
