+++ cmd/tailscale/depaware.txt
@@ -188,7 +188,7 @@ tailscale.com/cmd/tailscale dependencies: (generated by github.com/tailscale/dep
         tailscale.com/cmd/tailscale/cli/ffcomplete/internal          from tailscale.com/cmd/tailscale/cli/ffcomplete
         tailscale.com/cmd/tailscale/jsonoutput                       from tailscale.com/cmd/tailscale/cli+
         tailscale.com/cmd/tailscale/tsdnsjsonv0                      from tailscale.com/cmd/tailscale/cli
-        tailscale.com/cmd/tailscale/tslockjsonv1                     from tailscale.com/cmd/tailscale/cli
+        tailscale.com/cmd/tailscale/tslockjsonv1                     from tailscale.com/feature/tailnetlock/tslockjsonv1
         tailscale.com/cmd/tailscale/tsroutecheckjsonv0               from tailscale.com/cmd/tailscale/cli
         tailscale.com/control/controlbase                            from tailscale.com/control/controlhttp+
         tailscale.com/control/controlhttp                            from tailscale.com/control/ts2021
@@ -215,6 +215,7 @@ tailscale.com/cmd/tailscale dependencies: (generated by github.com/tailscale/dep
         tailscale.com/feature/portmapper                             from tailscale.com/feature/condregister/portmapper
         tailscale.com/feature/serviceclientprefs/serviceclient       from tailscale.com/client/local
         tailscale.com/feature/syspolicy                              from tailscale.com/cmd/tailscale/cli
+        tailscale.com/feature/tailnetlock/tslockjsonv1               from tailscale.com/cmd/tailscale/cli
         tailscale.com/feature/useproxy                               from tailscale.com/feature/condregister/useproxy
         tailscale.com/gokrazy/mkfs                                   from tailscale.com/cmd/tailscale/cli
         tailscale.com/health                                         from tailscale.com/net/tlsdial+
