+++ cmd/tailscale/tslockjsonv1/tailnet-lock-status.go
@@ -5,86 +5,7 @@
 
 package tslockjsonv1
 
-import (
-	jsonv1 "encoding/json"
-	"io"
-
-	"tailscale.com/cmd/tailscale/jsonoutput"
-	"tailscale.com/ipn/ipnstate"
-	"tailscale.com/tka"
-)
-
-// PrintTailnetLockStatusJSONV1 prints the current Tailnet Lock status
-// as a JSON object to the CLI, in a stable "v1" format.
-func PrintTailnetLockStatusJSONV1(out io.Writer, status *ipnstate.TailnetLockStatus) error {
-	responseEnvelope := jsonoutput.ResponseEnvelope{
-		SchemaVersion: "1",
-	}
-
-	var result StatusResponse
-	if status.Enabled {
-		result = toTailnetLockEnabledStatusV1(status)
-		result.ResponseEnvelope = responseEnvelope
-	} else {
-		result = toTailnetLockDisabledStatusV1(status)
-		result.ResponseEnvelope = responseEnvelope
-	}
-
-	enc := jsonv1.NewEncoder(out)
-	enc.SetIndent("", "  ")
-	return enc.Encode(result)
-}
-
-func toTailnetLockDisabledStatusV1(status *ipnstate.TailnetLockStatus) StatusResponse {
-	out := StatusResponse{
-		Enabled: status.Enabled,
-	}
-	if !status.PublicKey.IsZero() {
-		out.PublicKey = status.PublicKey.CLIString()
-	}
-	if nk := status.NodeKey; nk != nil {
-		out.NodeKey = nk.String()
-	}
-	return out
-}
-
-func toTailnetLockEnabledStatusV1(status *ipnstate.TailnetLockStatus) StatusResponse {
-	out := StatusResponse{
-		Enabled: status.Enabled,
-	}
-
-	if status.Head != nil {
-		var head tka.AUMHash
-		h := status.Head
-		copy(head[:], h[:])
-		out.Head = head.String()
-	}
-	if !status.PublicKey.IsZero() {
-		out.PublicKey = status.PublicKey.CLIString()
-	}
-	if nk := status.NodeKey; nk != nil {
-		out.NodeKey = nk.String()
-	}
-	out.NodeKeySigned = &status.NodeKeySigned
-	if sig := status.NodeKeySignature; sig != nil {
-		out.NodeKeySignature = toTKANodeKeySignatureV1(sig)
-	}
-	out.TrustedKeys = []Key{} // never omit this field when enabled
-	for _, key := range status.TrustedKeys {
-		out.TrustedKeys = append(out.TrustedKeys, ipnTKAKeytoTKAKeyV1(&key))
-	}
-	out.VisiblePeers = []TrustedPeer{} // never omit this field when enabled
-	for _, vp := range status.VisiblePeers {
-		out.VisiblePeers = append(out.VisiblePeers, toTrustedTKAPeerV1(vp))
-	}
-	out.FilteredPeers = []Peer{} // never omit this field when enabled
-	for _, fp := range status.FilteredPeers {
-		out.FilteredPeers = append(out.FilteredPeers, toTKAPeerV1(fp))
-	}
-	out.StateID = status.StateID
-
-	return out
-}
+import "tailscale.com/cmd/tailscale/jsonoutput"
 
 // StatusResponse is the full Tailnet Lock Status output collected from the local Tailscale daemon.
 type StatusResponse struct {
