+++ cmd/tailscaled/depaware.txt
@@ -267,8 +267,8 @@ tailscale.com/cmd/tailscaled dependencies: (generated by github.com/tailscale/de
         tailscale.com/client/web                                     from tailscale.com/ipn/ipnlocal
         tailscale.com/clientupdate                                   from tailscale.com/feature/clientupdate
   LW    tailscale.com/clientupdate/distsign                          from tailscale.com/clientupdate
-        tailscale.com/cmd/tailscale/jsonoutput                       from tailscale.com/cmd/tailscale/tslockjsonv1
-        tailscale.com/cmd/tailscale/tslockjsonv1                     from tailscale.com/feature/tailnetlock
+        tailscale.com/cmd/tailscale/jsonoutput                       from tailscale.com/cmd/tailscale/tslockjsonv1+
+        tailscale.com/cmd/tailscale/tslockjsonv1                     from tailscale.com/feature/tailnetlock/tslockjsonv1
         tailscale.com/cmd/tailscaled/childproc                       from tailscale.com/cmd/tailscaled+
         tailscale.com/cmd/tailscaled/tailscaledhooks                 from tailscale.com/cmd/tailscaled+
         tailscale.com/control/controlbase                            from tailscale.com/control/controlhttp+
@@ -330,6 +330,7 @@ tailscale.com/cmd/tailscaled dependencies: (generated by github.com/tailscale/de
         tailscale.com/feature/syspolicy                              from tailscale.com/feature/condregister
         tailscale.com/feature/taildrop                               from tailscale.com/feature/condregister
         tailscale.com/feature/tailnetlock                            from tailscale.com/feature/condregister
+        tailscale.com/feature/tailnetlock/tslockjsonv1               from tailscale.com/feature/tailnetlock
    L    tailscale.com/feature/tap                                    from tailscale.com/feature/condregister
         tailscale.com/feature/tpm                                    from tailscale.com/feature/condregister
    L 💣 tailscale.com/feature/tundevstats                            from tailscale.com/feature/condregister
