+++ feature/tailnetlock/tslockjsonv1/doc.go
@@ -1,0 +1,9 @@
+// Copyright (c) Tailscale Inc & contributors
+// SPDX-License-Identifier: BSD-3-Clause
+
+//go:build !ts_omit_tailnetlock
+
+// Package tslockjsonv1 provides converters for internal data structures
+// to the JSON output returned by the CLI, in a stable "v1" format,
+// defined in the [tailscale.com/cmd/tailscale/tslockjsonv1] package.
+package tslockjsonv1
